Refactory archive package to use interfaces
See #20
