+++ internal/archive/ofb_envelop_test.go
@@ -1,281 +1,24 @@
 package archive_test
 
 import (
-	"archive/tar"
 	"crypto/aes"
 	"crypto/rand"
 	"encoding/hex"
 	"errors"
-	"fmt"
 	"io"
 	"io/ioutil"
 	"os"
 	"path"
 	"reflect"
-	"regexp"
 	"testing"
 
 	"github.com/rafaeljusto/toglacier/internal/archive"
 )
 
-func TestBuild(t *testing.T) {
-	scenarios := []struct {
-		description   string
-		backupPaths   []string
-		expected      func(filename string) error
-		expectedError error
-	}{
-		{
-			description: "it should create an archive correctly from directory path",
-			backupPaths: func() []string {
-				d, err := ioutil.TempDir("", "toglacier-test")
-				if err != nil {
-					t.Fatalf("error creating temporary directory. details %s", err)
-				}
-
-				if err := ioutil.WriteFile(path.Join(d, "file1"), []byte("file1 test"), os.ModePerm); err != nil {
-					t.Fatalf("error creating temporary file. details %s", err)
-				}
-
-				if err := ioutil.WriteFile(path.Join(d, "file2"), []byte("file2 test"), os.ModePerm); err != nil {
-					t.Fatalf("error creating temporary file. details %s", err)
-				}
-
-				if err := os.Mkdir(path.Join(d, "dir1"), os.ModePerm); err != nil {
-					t.Fatalf("error creating temporary directory. details %s", err)
-				}
-
-				if err := ioutil.WriteFile(path.Join(d, "dir1", "file3"), []byte("file3 test"), os.ModePerm); err != nil {
-					t.Fatalf("error creating temporary file. details %s", err)
-				}
-
-				// add an empty directory to see if it's going to be ignored
-				return []string{d, ""}
-			}(),
-			expected: func(filename string) error {
-				f, err := os.Open(filename)
-				if err != nil {
-					return fmt.Errorf("error opening archive. details: %s", err)
-				}
-				defer f.Close()
-
-				basePath := `backup-[0-9]+`
-				expectedFiles := []*regexp.Regexp{
-					regexp.MustCompile(`^` + basePath + `/$`),
-					regexp.MustCompile(`^` + path.Join(basePath, `file1`) + `$`),
-					regexp.MustCompile(`^` + path.Join(basePath, `file2`) + `$`),
-					regexp.MustCompile(`^` + path.Join(basePath, `dir1`) + `/$`),
-					regexp.MustCompile(`^` + path.Join(basePath, `dir1`, `file3`) + `$`),
-				}
-
-				tr := tar.NewReader(f)
-				for {
-					hdr, err := tr.Next()
-					if err == io.EOF {
-						break
-					} else if err != nil {
-						return err
-					}
-
-					if len(expectedFiles) == 0 {
-						return fmt.Errorf("content “%s” shouldn't be here", hdr.Name)
-					}
-
-					found := false
-					for i, expectedFile := range expectedFiles {
-						if expectedFile.MatchString(hdr.Name) {
-							expectedFiles = append(expectedFiles[:i], expectedFiles[i+1:]...)
-							found = true
-							break
-						}
-					}
-
-					if !found {
-						return fmt.Errorf("file “%s” did not match with any of the expected files", hdr.Name)
-					}
-				}
-
-				if len(expectedFiles) > 0 {
-					return errors.New("not all files were found in the archive")
-				}
-
-				return nil
-			},
-		},
-		{
-			description: "it should create an archive correctly from file path",
-			backupPaths: func() []string {
-				f, err := ioutil.TempFile("", "toglacier-test")
-				if err != nil {
-					t.Fatalf("error creating temporary directory. details %s", err)
-				}
-				defer f.Close()
-
-				f.WriteString("file test")
-
-				return []string{f.Name()}
-			}(),
-			expected: func(filename string) error {
-				f, err := os.Open(filename)
-				if err != nil {
-					return fmt.Errorf("error opening archive. details: %s", err)
-				}
-				defer f.Close()
-
-				expectedFiles := []*regexp.Regexp{
-					regexp.MustCompile(`^toglacier-test[0-9]+$`),
-				}
-
-				tr := tar.NewReader(f)
-				for {
-					hdr, err := tr.Next()
-					if err == io.EOF {
-						break
-					} else if err != nil {
-						return err
-					}
-
-					if len(expectedFiles) == 0 {
-						return fmt.Errorf("content “%s” shouldn't be here", hdr.Name)
-					}
-
-					found := false
-					for i, expectedFile := range expectedFiles {
-						if expectedFile.MatchString(hdr.Name) {
-							expectedFiles = append(expectedFiles[:i], expectedFiles[i+1:]...)
-							found = true
-							break
-						}
-					}
-
-					if !found {
-						return fmt.Errorf("file “%s” did not match with any of the expected files", hdr.Name)
-					}
-				}
-
-				if len(expectedFiles) > 0 {
-					return errors.New("not all files were found in the archive")
-				}
-
-				return nil
-			},
-		},
-		{
-			description: "it should detect when the path does not exist",
-			backupPaths: func() []string {
-				return []string{"idontexist12345"}
-			}(),
-			expectedError: &archive.PathError{
-				Path: "idontexist12345",
-				Code: archive.PathErrorCodeInfo,
-				Err: &os.PathError{
-					Op:   "lstat",
-					Path: "idontexist12345",
-					Err:  errors.New("no such file or directory"),
-				},
-			},
-		},
-		{
-			description: "it should detect when the path (directory) does not have permission",
-			backupPaths: func() []string {
-				n := path.Join(os.TempDir(), "toglacier-test-archive-dir-noperm")
-				if _, err := os.Stat(n); os.IsNotExist(err) {
-					err := os.Mkdir(n, os.FileMode(0077))
-					if err != nil {
-						t.Fatalf("error creating a temporary directory. details: %s", err)
-					}
-				}
-
-				return []string{n}
-			}(),
-			expectedError: &archive.PathError{
-				Path: path.Join(os.TempDir(), "toglacier-test-archive-dir-noperm"),
-				Code: archive.PathErrorCodeInfo,
-				Err: &os.PathError{
-					Op:   "open",
-					Path: path.Join(os.TempDir(), "toglacier-test-archive-dir-noperm"),
-					Err:  errors.New("permission denied"),
-				},
-			},
-		},
-		{
-			description: "it should detect when the path (file) does not have permission",
-			backupPaths: func() []string {
-				n := path.Join(os.TempDir(), "toglacier-test-archive-file-noperm")
-				if _, err := os.Stat(n); os.IsNotExist(err) {
-					f, err := os.OpenFile(n, os.O_CREATE, os.FileMode(0077))
-					if err != nil {
-						t.Fatalf("error creating a temporary file. details: %s", err)
-					}
-					defer f.Close()
-
-					f.WriteString("This is a test")
-				}
-
-				return []string{n}
-			}(),
-			expectedError: &archive.PathError{
-				Path: path.Join(os.TempDir(), "toglacier-test-archive-file-noperm"),
-				Code: archive.PathErrorCodeOpeningFile,
-				Err: &os.PathError{
-					Op:   "open",
-					Path: path.Join(os.TempDir(), "toglacier-test-archive-file-noperm"),
-					Err:  errors.New("permission denied"),
-				},
-			},
-		},
-		{
-			description: "it should detect an error while walking in the path",
-			backupPaths: func() []string {
-				n := path.Join(os.TempDir(), "toglacier-test-archive-dir-file-noperm")
-				if _, err := os.Stat(n); os.IsNotExist(err) {
-					err := os.Mkdir(n, os.FileMode(0700))
-					if err != nil {
-						t.Fatalf("error creating a temporary directory. details: %s", err)
-					}
-
-					f, err := os.OpenFile(path.Join(n, "file1"), os.O_CREATE, os.FileMode(0077))
-					if err != nil {
-						t.Fatalf("error creating a temporary file. details: %s", err)
-					}
-					defer f.Close()
-
-					f.WriteString("file1 test")
-				}
-
-				return []string{n}
-			}(),
-			expectedError: &archive.PathError{
-				Path: path.Join(os.TempDir(), "toglacier-test-archive-dir-file-noperm", "file1"),
-				Code: archive.PathErrorCodeOpeningFile,
-				Err: &os.PathError{
-					Op:   "open",
-					Path: path.Join(os.TempDir(), "toglacier-test-archive-dir-file-noperm", "file1"),
-					Err:  errors.New("permission denied"),
-				},
-			},
-		},
-	}
-
-	for _, scenario := range scenarios {
-		t.Run(scenario.description, func(t *testing.T) {
-			filename, err := archive.Build(scenario.backupPaths...)
-			if scenario.expectedError == nil && scenario.expected != nil {
-				if err := scenario.expected(filename); err != nil {
-					t.Errorf("unexpected archive content (%s). details: %s", filename, err)
-				}
-			}
-
-			if !archive.ErrorEqual(scenario.expectedError, err) && !archive.PathErrorEqual(scenario.expectedError, err) {
-				t.Errorf("errors don't match. expected “%v” and got “%v”", scenario.expectedError, err)
-			}
-		})
-	}
-}
-
-func TestEncrypt(t *testing.T) {
+func TestOFBEnvelop_Encrypt(t *testing.T) {
 	type scenario struct {
 		description   string
+		envelop       archive.OFBEnvelop
 		filename      string
 		secret        string
 		randomSource  io.Reader
@@ -386,7 +129,7 @@ func TestEncrypt(t *testing.T) {
 	for _, scenario := range scenarios {
 		t.Run(scenario.description, func(t *testing.T) {
 			archive.RandomSource = scenario.randomSource
-			encryptedFilename, err := archive.Encrypt(scenario.filename, scenario.secret)
+			encryptedFilename, err := scenario.envelop.Encrypt(scenario.filename, scenario.secret)
 
 			fileContent, fileErr := ioutil.ReadFile(encryptedFilename)
 			if fileErr != nil && scenario.expectedError == nil {
@@ -404,9 +147,10 @@ func TestEncrypt(t *testing.T) {
 	}
 }
 
-func TestDecrypt(t *testing.T) {
+func TestOFBEnvelop_Decrypt(t *testing.T) {
 	type scenario struct {
 		description       string
+		envelop           archive.OFBEnvelop
 		encryptedFilename string
 		secret            string
 		expectedFile      string
@@ -523,7 +267,7 @@ func TestDecrypt(t *testing.T) {
 
 	for _, scenario := range scenarios {
 		t.Run(scenario.description, func(t *testing.T) {
-			filename, err := archive.Decrypt(scenario.encryptedFilename, scenario.secret)
+			filename, err := scenario.envelop.Decrypt(scenario.encryptedFilename, scenario.secret)
 
 			fileContent, fileErr := ioutil.ReadFile(filename)
 			if fileErr != nil && scenario.expectedError == nil {
@@ -541,9 +285,10 @@ func TestDecrypt(t *testing.T) {
 	}
 }
 
-func Test_EncryptDecrypt(t *testing.T) {
+func TestOFBEnvelop_EncryptDecrypt(t *testing.T) {
 	scenarios := []struct {
 		description          string
+		envelop              archive.OFBEnvelop
 		filename             string
 		secret               string
 		expectedFile         string
@@ -582,7 +327,7 @@ func Test_EncryptDecrypt(t *testing.T) {
 
 	for _, scenario := range scenarios {
 		t.Run(scenario.description, func(t *testing.T) {
-			encryptedFilename, err := archive.Encrypt(scenario.filename, scenario.secret)
+			encryptedFilename, err := scenario.envelop.Encrypt(scenario.filename, scenario.secret)
 			if !reflect.DeepEqual(scenario.expectedEncryptError, err) {
 				t.Fatalf("errors don't match. expected “%v” and got “%v”", scenario.expectedEncryptError, err)
 			}
@@ -591,7 +336,7 @@ func Test_EncryptDecrypt(t *testing.T) {
 				return
 			}
 
-			filename, err := archive.Decrypt(encryptedFilename, scenario.secret)
+			filename, err := scenario.envelop.Decrypt(encryptedFilename, scenario.secret)
 			if !reflect.DeepEqual(scenario.expectedDecryptError, err) {
 				t.Errorf("errors don't match. expected “%v” and got “%v”", scenario.expectedDecryptError, err)
 			}
